refactor: extract WorkItem domain types
