+++ src/lib.rs
@@ -12,6 +12,7 @@ pub mod config;
 pub mod context;
 pub mod daemon;
 pub mod diagnostics;
+pub mod domain;
 pub mod fd_limit;
 pub mod host;
 mod host_registry;
